requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
realesrgan==0.3.1
insightface==0.7.3
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
torch==2.2.0
torchvision==0.17.0